refactor: co-locate audit session and preference state
