+++ features/degree-audit-app/audit-view/requirement-breakdown.tsx
@@ -14,7 +14,7 @@ import {
 import { useState } from "react";
 import EyeIcon from "@/assets/svgs/Eye.svg";
 import { useCourseModalContext } from "../course-search/course-modal-provider";
-import { useAuditContext } from "../providers/audit-provider";
+import { useAuditContext } from "@/features/audit/audit-provider";
 
 type RequirementCompletionState = "completed" | "not-started" | "in-progress";
 type ProgressLabelUnit = RequirementProgressUnit | "progress";
