+++ features/degree-audit-app/planner-view/planner-course-card.tsx
@@ -2,7 +2,7 @@ import type { CourseId } from "@/domain/course";
 import { cn } from "@/lib/utils";
 import { useDraggable } from "@dnd-kit/core";
 import CourseCard from "../course-search/course-card";
-import { useAuditContext } from "../providers/audit-provider";
+import { useAuditContext } from "@/features/audit/audit-provider";
 
 interface PlannerCourseCardProps {
   courseId: CourseId;
